20220228.0.BETA
Adds the optional `tempdir` config option to specify where temp files should go. Most users will not need this but may be useful to try to avoid writes on some drives. Note, this is **independent** of rclone's temp directoru settings (see `--temp-dir` in [the docs]) *and* there is very little that needs to be written! Closes #18 and closes #20

Copy from local devel repo: 2022-02-28 16:30:38 -0700|9f36ca4